5.2.1 Перенесет функции для рендеринга из  в
